add constants to better id the buttons
By just using the array index to identify which button was which, it was
hard to determine the button upon visual inpection of the code.  To
address this, a few constants were introduced to effectively label each
button in the JButton array.  This commit also fixes some whitespace and
formatting issues.

* add contants to better identify the JButtons in the array
* fixes whitespace and formatting issues.